% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn-to-sql.R
\name{pgn2sql}
\alias{pgn2sql}
\title{Convert a PGN file to a SQLite DB}
\usage{
pgn2sql(paths, sql_out)
}
\arguments{
\item{paths}{path(s) to PGN file(s)}

\item{sql_out}{SQLite database file (will be created if it doesn't exist)}
}
\description{
Read one ore more PGN files and store the game(s) into a SQLite database under
a \code{games} table (which will be created if it does not exist).
}
\details{
A \code{.} will be printed to stdout for ever 1,000 records inserted.
}
\examples{
tf <- tempfile(fileext = ".db")
pgn2sql(system.file("extdata", "r7.pgn", package="pigeon"), tf)

tmp <- dplyr::src_sqlite(tf)
dplyr::glimpse(dplyr::tbl(tmp, "games"))

unlink(tf)
}
