% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn-count.R
\name{pgn_count}
\alias{pgn_count}
\title{Count number of games in a PGN file}
\usage{
pgn_count(path, n = Inf, sample = FALSE)
}
\arguments{
\item{path}{path to PGN file}
}
\value{
numeric (count of gamesz in PGN file)
}
\description{
Count number of games in a PGN file
}
