% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{pgn_parse_moves}
\alias{pgn_parse_moves}
\title{Parse a character vector of chess moves in algebraic notations}
\usage{
pgn_parse_moves(moves)
}
\arguments{
\item{moves}{character vector of chess moves}
}
\value{
character vector of the invidivual components of `moves`
}
\description{
Parse a character vector of chess moves in algebraic notations
}
