% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigeon.r
\name{read_pgn}
\alias{read_pgn}
\title{Read in a PGN file}
\usage{
read_pgn(path)
}
\arguments{
\item{path}{path to PGN file}
}
\value{
data frame of results or an empty data frame if parsing errors occurred
}
\description{
Read in a PGN file
}
\examples{
read_pgn(system.file("extdata", "r7.pgn", package="pigeon"))
}
