% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn-to-sql.R
\name{pgn2sql}
\alias{pgn2sql}
\title{Convert a PGN file to a SQLite DB}
\usage{
pgn2sql(paths, sql_out, overwrite = TRUE)
}
\arguments{
\item{paths}{path(s) to PGN file(s)}

\item{sql_out}{SQLite database file (will be created if it doesn't exist)}

\item{overwrite}{overwrite the \code{games} table if it exists?}
}
\description{
Read one ore more PGN files and store the game(s) into a SQLite database.
}
\details{
If \code{sql_out} exists and \code{overwrite} is \code{TRUE} then the \code{games} table will be
dropped and overwritten. If \code{overwrite} if \code{FALSE} and \code{sql_out} exists and
there is a \code{games} table, the \code{games} table will be appended to.
}
