% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pgn2.R
\name{read_pgn}
\alias{read_pgn}
\title{Read chess match files in Portable Game Notation (PGN) format}
\usage{
read_pgn(path, n = Inf, sample = FALSE)
}
\arguments{
\item{path}{path to PGN file}

\item{n}{if not \code{Inf} then the number of games to read in. If \code{n} > total
games in the PGN file, it will only read in that may games. This is
a sequential operation. If you want this to be a random sample,
set \code{sample} to \code{TRUE} and \code{n} to be the number of samples you want.}

\item{sample}{if \code{TRUE} \strong{and} \code{n} is \strong{not} \code{Inf} then the \code{n} games
retrieved will be from a random sample. Use \code{set.seed()} before
calling this if you want the results to be reproducible.}
}
\description{
Read chess match files in Portable Game Notation (PGN) format
}
